build: Add Apollo client
