+++ gatsby-browser.js
@@ -1,23 +1,5 @@
 import * as React from 'react';
 import { wrapPageElement as wrap } from './src/chakra-wrapper';
-import { ApolloWrapper } from './src/apollo/apollo-wrapper';
-import { Auth0Provider } from '@auth0/auth0-react';
-import { navigate } from 'gatsby';
 
-const onRedirectCallback = (appState) => {
-  // Use Gatsby's navigate method to replace the url
-  navigate(appState?.returnTo || '/', { replace: true });
-};
-
-export const wrapRootElement = ({ element }) => (
-  <Auth0Provider
-    domain={process.env.GATSBY_AUTH0_DOMAIN}
-    clientId={process.env.GATSBY_AUTH0_CLIENT_ID}
-    redirectUri={window.location.origin}
-    onRedirectCallback={onRedirectCallback}
-  >
-    {element}
-  </Auth0Provider>
-);
-
+export { wrapRootElement } from './gatsby-ssr';
 export const wrapPageElement = wrap;
