+++ src/apollo/apollo-wrapper.js
@@ -2,6 +2,6 @@ import React from 'react';
 import { ApolloProvider } from '@apollo/client';
 import { client } from './client';
 
-export const ApolloWrapper = ({ element }) => (
-  <ApolloProvider client={client}>{element}</ApolloProvider>
+export const ApolloWrapper = (props) => (
+  <ApolloProvider client={client}>{props.children}</ApolloProvider>
 );
